requirements.txt
google-api-python-client==2.111.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
pandas==2.1.4
requests==2.31.0
